Extract BaseFrontierMode: eliminate ~120 lines of duplication
StandardMode and CircleCoopMode shared identical frontier logic:
handleFrontierAction, handleFrontierBatchAction, rotateRandomFrontier,
frontier panel setup, wave clear bonus collection.

New BaseFrontierMode abstract class contains all shared code:
- StandardMode extends it, adding only SendPanel logic (~79 lines, was ~180)
- CircleCoopMode extends it with zero overrides (~10 lines, was ~141)

GameScene uses instanceof BaseFrontierMode for Random rotation,
so both modes get frontier rotation correctly.
